+++ package.json
@@ -31,31 +31,33 @@
     "@lukemorales/query-key-factory": "^1.3.4",
     "@radix-ui/react-dialog": "^1.1.1",
     "@radix-ui/react-slot": "^1.1.0",
-    "@sentry/react": "^8.4.0",
-    "@sentry/vite-plugin": "^2.17.0",
-    "@tanstack/react-query": "^5.39.0",
-    "@tanstack/react-query-devtools": "^5.39.0",
+    "@sentry/integrations": "^7.114.0",
+    "@sentry/react": "^8.30.0",
+    "@sentry/tracing": "^7.114.0",
+    "@sentry/vite-plugin": "^2.22.4",
+    "@tanstack/react-query": "^5.56.2",
+    "@tanstack/react-query-devtools": "^5.56.2",
     "axios": "^1.7.2",
     "date-fns": "^3.6.0",
     "jwt-decode": "^4.0.0",
     "mem": "^10.0.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-hook-form": "^7.51.5",
-    "react-router-dom": "^6.23.1",
-    "tailwind-merge": "^2.3.0",
+    "react-hook-form": "^7.53.0",
+    "react-router-dom": "^6.26.2",
+    "tailwind-merge": "^2.5.2",
     "tailwind-variants": "^0.2.1",
-    "tailwindcss": "3.4.3",
+    "tailwindcss": "3.4.12",
     "tailwindcss-animate": "^1.0.7",
-    "uuid": "^9.0.1",
+    "uuid": "^10.0.0",
     "zod": "^3.23.8",
-    "zustand": "^4.5.2"
+    "zustand": "^4.5.5"
   },
   "devDependencies": {
     "@eslint/compat": "^1.1.1",
     "@eslint/js": "^9.8.0",
     "@ianvs/prettier-plugin-sort-imports": "^4.1.1",
-    "@sentry/types": "^8.29.0",
+    "@sentry/types": "^8.30.0",
     "@storybook/addon-a11y": "^7.6.17",
     "@storybook/addon-actions": "^7.6.17",
     "@storybook/addon-essentials": "^7.6.17",
@@ -89,7 +91,7 @@
     "eslint-plugin-react": "^7.35.0",
     "eslint-plugin-react-hooks": "^4.6.2",
     "globals": "^15.8.0",
-    "husky": "^9.0.11",
+    "husky": "^9.1.6",
     "lint-staged": "^15.2.2",
     "plop": "^4.0.1",
     "postcss": "^8.4.35",
@@ -98,9 +100,9 @@
     "pretty-quick": "^4.0.0",
     "storybook": "^7.6.17",
     "ts-node": "^10.9.2",
-    "typescript": "^5.4.2",
-    "typescript-eslint": "^7.17.0",
-    "vite": "^5.1.7",
+    "typescript": "^5.6.2",
+    "typescript-eslint": "^8.6.0",
+    "vite": "^5.4.6",
     "yargs": "^17.7.2"
   },
   "postcss": {
