+++ src/domains/index.ts
@@ -1,1 +1,2 @@
 export * from "./guest";
+export * from "./medications";
