+++ src/providers/SentryProvider.tsx
@@ -1,0 +1,13 @@
+import type { PropsWithChildren } from "react";
+import React from "react";
+import * as Sentry from "@sentry/react";
+
+import { ErrorBoundary } from "~/sections";
+
+export const SentryProvider = ({ children }: PropsWithChildren) => {
+  return (
+    <Sentry.ErrorBoundary fallback={<ErrorBoundary />}>
+      {children}
+    </Sentry.ErrorBoundary>
+  );
+};
