--- src/shared/index.ts
@@ -1,3 +1,0 @@
-export * from "./hooks";
-export * from "./components";
-export * from "./utils";
